feat: Add AudioPlayer component to PlayerView and TableView, and remove unused video elements
