+++ ui/src/components/PlayerFileList.vue
@@ -15,38 +15,27 @@
       <tbody>
         <tr v-for="track in audioStore.availableTracks" :key="track.fileName">
           <td>{{ track.fileName }}</td>
-          <td class="d-flex align-center">
-            <video :ref="el => videoElements[track.fileName] = el as HTMLVideoElement" />
-          </td>
+          <td class="d-flex align-center"></td>
         </tr>
       </tbody>
     </v-table>
+    <AudioPlayer v-if="audioStore.enabled" />
   </v-container>
 </template>
 
 <script setup lang="ts">
 import { wsHandlers } from '@/composables/wsHandlers'
 import { useWebSocketStore } from '@/stores/websocket'
-import { onBeforeUnmount, ref, watch } from 'vue'
-import { useAudioSync } from '../composables/useAudioSync'
+import { ref } from 'vue'
 import { useAudioStore } from '../stores/audio'
+import AudioPlayer from './AudioPlayer.vue'
 
 const audioStore = useAudioStore()
-const videoElements = ref<Record<string, HTMLVideoElement>>({})
 const wsStore = useWebSocketStore()
 const isRefreshing = ref(false)
 
 // Set up sync handling with audio elements
-wsHandlers(videoElements)
-
-// Set up audio sync for new elements
-watch(videoElements, (elements) => {
-  Object.entries(elements).forEach(([fileName, video]) => {
-    if (video) {
-      useAudioSync(fileName, video)
-    }
-  })
-}, { deep: true })
+wsHandlers()
 
 function handleRefresh() {
   isRefreshing.value = true
@@ -55,13 +44,6 @@ function handleRefresh() {
     isRefreshing.value = false
   }, 1000)
 }
-
-onBeforeUnmount(() => {
-  Object.values(videoElements.value).forEach(video => {
-    video.pause()
-    video.src = ''
-  })
-})
 </script>
 
 <style scoped>
